requirements.txt
streamlit>=1.24.0
langchain>=0.0.300
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.38.0
requests>=2.31.0
python-dotenv>=1.0.0
pymupdf>=1.22.0  
tqdm
